modify spider for run on platform

+++ spider/src/main/java/com/maplecloudy/spider/crawl/Injector.java
@@ -10,7 +10,6 @@ import java.util.Random;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.Text;
@@ -22,12 +21,13 @@ import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
 import org.apache.hadoop.util.StringUtils;
 import org.apache.hadoop.util.Tool;
-import org.apache.hadoop.util.ToolRunner;
 
 import com.maplecloudy.avro.mapreduce.AvroJob;
 import com.maplecloudy.avro.mapreduce.input.AvroPairInputFormat;
 import com.maplecloudy.avro.mapreduce.output.AvroMapOutputFormat;
 import com.maplecloudy.avro.mapreduce.output.AvroPairOutputFormat;
+import com.maplecloudy.oozie.main.OozieMain;
+import com.maplecloudy.oozie.main.OozieToolRunner;
 import com.maplecloudy.spider.net.BasicURLNormalizer;
 import com.maplecloudy.spider.util.SpiderConfiguration;
 
@@ -43,26 +43,23 @@ import com.maplecloudy.spider.util.SpiderConfiguration;
  * e.g. http://www.nutch.org/ \t nutch.score=10 \t nutch.fetchInterval=2592000
  * \t userType=open_source
  **/
-public class Injector extends Configured implements Tool {
-	public static final Log	LOG	= LogFactory.getLog(Injector.class);
+public class Injector extends OozieMain implements Tool {
+	public static final Log LOG = LogFactory.getLog(Injector.class);
 
 	/** Normalize and filter injected urls. */
-	public static class InjectMapper extends
-			Mapper<WritableComparable, Text, String, CrawlDatum> {
+	public static class InjectMapper extends Mapper<WritableComparable, Text, String, CrawlDatum> {
 
-		private int									interval;
-		private float								scoreInjected;
-		private Configuration				jobConf;
-		private long								curTime;
+		private int interval;
+		private float scoreInjected;
+		private Configuration jobConf;
+		private long curTime;
 
 		@Override
-		protected void setup(Context context) throws IOException,
-				InterruptedException {
+		protected void setup(Context context) throws IOException, InterruptedException {
 			jobConf = context.getConfiguration();
 			interval = jobConf.getInt("db.fetch.interval.default", 2592000);
 			scoreInjected = jobConf.getFloat("db.score.injected", 1.0f);
-			curTime = jobConf.getLong("injector.current.time",
-					System.currentTimeMillis());
+			curTime = jobConf.getLong("injector.current.time", System.currentTimeMillis());
 		}
 
 		@Override
@@ -98,8 +95,7 @@ public class Injector extends Configured implements Tool {
 					url = null;
 				}
 				if (url != null) { // if it passes
-					CrawlDatum datum = new CrawlDatum(CrawlDatum.STATUS_INJECTED,
-							customInterval);
+					CrawlDatum datum = new CrawlDatum(CrawlDatum.STATUS_INJECTED, customInterval);
 					datum.setFetchTime(curTime);
 					if (customScore != -1)
 						datum.setScore(customScore);
@@ -116,18 +112,17 @@ public class Injector extends Configured implements Tool {
 	}
 
 	/** Combine multiple new entries for a url. */
-	public static class InjectReducer extends
-			Reducer<String, CrawlDatum, String, CrawlDatum> {
+	public static class InjectReducer extends Reducer<String, CrawlDatum, String, CrawlDatum> {
 
-		private CrawlDatum	old				= new CrawlDatum();
-		private CrawlDatum	injected	= new CrawlDatum();
+		private CrawlDatum old = new CrawlDatum();
+		private CrawlDatum injected = new CrawlDatum();
 
 		@Override
 		protected void reduce(String key, Iterable<CrawlDatum> values, Context context)
 				throws IOException, InterruptedException {
 			boolean oldSet = false;
 			for (CrawlDatum cur : values) {
-				
+
 				if (cur.getStatus() == CrawlDatum.STATUS_INJECTED) {
 					injected.set(cur);
 					injected.setStatus(CrawlDatum.STATUS_DB_UNFETCHED);
@@ -152,7 +147,8 @@ public class Injector extends Configured implements Tool {
 		setConf(conf);
 	}
 
-	public void inject(Path crawlDb, Path urlDir) throws IOException, IllegalStateException, InterruptedException, ClassNotFoundException {
+	public int inject(Path crawlDb, Path urlDir)
+			throws IOException, ClassNotFoundException, IllegalStateException, InterruptedException {
 
 		if (LOG.isInfoEnabled()) {
 			LOG.info("Injector: starting");
@@ -160,15 +156,14 @@ public class Injector extends Configured implements Tool {
 			LOG.info("Injector: urlDir: " + urlDir);
 		}
 
-		Path tempDir = new Path(getConf().get("mapred.temp.dir", ".")
-				+ "/inject-temp-"
+		Path tempDir = new Path(getConf().get("mapred.temp.dir", ".") + "/inject-temp-"
 				+ Integer.toString(new Random().nextInt(Integer.MAX_VALUE)));
 
 		// map text input file to a <url,CrawlDatum> file
 		if (LOG.isInfoEnabled()) {
 			LOG.info("Injector: Converting injected urls to crawl db entries.");
 		}
-		Job sortJob = AvroJob.getAvroJob(getConf());
+		AvroJob sortJob = AvroJob.getAvroJob(getConf());
 		sortJob.setJobName("inject " + urlDir);
 		FileInputFormat.addInputPath(sortJob, urlDir);
 		sortJob.setMapperClass(InjectMapper.class);
@@ -177,12 +172,12 @@ public class Injector extends Configured implements Tool {
 		sortJob.setOutputValueClass(CrawlDatum.class);
 		sortJob.setOutputFormatClass(AvroPairOutputFormat.class);
 
-		if (sortJob.waitForCompletion(true)) {
+		if (runJob(sortJob)) {
 			// merge with existing crawl db
 			if (LOG.isInfoEnabled()) {
 				LOG.info("Injector: Merging injected urls into crawl db.");
 			}
-			Job job = AvroJob.getAvroJob(getConf());
+			AvroJob job = AvroJob.getAvroJob(getConf());
 			job.setJobName("crawldb " + crawlDb);
 
 			Path current = new Path(crawlDb, CrawlDb.CURRENT_NAME);
@@ -193,28 +188,30 @@ public class Injector extends Configured implements Tool {
 			job.setInputFormatClass(AvroPairInputFormat.class);
 			job.setReducerClass(InjectReducer.class);
 			job.setMapperClass(CrawlDbFilter.class);
-			Path newCrawlDb = new Path(crawlDb, Integer.toString(new Random()
-			.nextInt(Integer.MAX_VALUE)));
+			Path newCrawlDb = new Path(crawlDb, Integer.toString(new Random().nextInt(Integer.MAX_VALUE)));
 			FileOutputFormat.setOutputPath(job, newCrawlDb);
 			job.setOutputFormatClass(AvroMapOutputFormat.class);
 			job.setOutputKeyClass(String.class);
 			job.setOutputValueClass(CrawlDatum.class);
-			job.waitForCompletion(true);
-			
-			CrawlDb.install(job, crawlDb);
-
-			// clean up
-			FileSystem fs = FileSystem.get(getConf());
-			fs.delete(tempDir, true);
-			if (LOG.isInfoEnabled()) {
-				LOG.info("Injector: done");
+			if (runJob(job)) {
+				CrawlDb.install(job, crawlDb);
+				FileSystem fs = FileSystem.get(getConf());
+				fs.delete(tempDir, true);
+				if (LOG.isInfoEnabled()) {
+					LOG.info("Injector: done");
+				}
+				return 0;
+			} else {
+				return -1;
 			}
+
+		} else {
+			return -1;
 		}
 	}
 
 	public static void main(String[] args) throws Exception {
-		int res = ToolRunner
-				.run(SpiderConfiguration.create(), new Injector(), args);
+		int res = OozieToolRunner.run(SpiderConfiguration.create(), new Injector(), args);
 		System.exit(res);
 	}
